Feat (BE-036): refactor create collection endpoint
